requirements.txt
Flask==3.0.0
firebase-admin==6.2.0
authlib==1.2.1
requests==2.31.0
python-dotenv==1.0.0
